Use new sci interface
